Reusable bitcoin jsonrpc network serialization
Pull the serialization code into the ledger module so we can reuse it
in Nectar.
