+++ comit/Cargo.toml
@@ -44,7 +44,7 @@ uuid = { version = "0.8", features = ["serde", "v4"] }
 [dev-dependencies]
 atty = "0.2"
 bitcoincore-rpc = "0.11"
-libp2p = { version = "0.28", default-features = false, features = ["yamux"] }
+libp2p = { version = "0.28", default-features = false, features = ["yamux", "noise"] }
 log = { version = "0.4", features = ["serde"] }
 proptest = "0.10"
 spectral = { version = "0.6", default-features = false }
